backend/requirements.txt
# DeepView Backend Dependencies
# Core FastAPI framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# AI/ML libraries
transformers>=4.35.0
torch>=2.0.0
numpy>=1.24.0

# Vector search
faiss-cpu>=1.7.4
# faiss-gpu>=1.7.4  # Uncomment for GPU support

# Data validation
pydantic>=2.0.0

# CORS middleware
fastapi-cors>=0.0.6

# Optional: For production deployment
# gunicorn>=21.0.0
# python-multipart>=0.0.6 